requirements.txt
httplib2==0.20.1
google-auth-oauthlib==0.4.6
google-auth==1.35.0
google-api-python-client==2.21.0
oauth2client==4.1.3
python-dateutil==2.8.2
openpyxl==3.0.9
pytz==2021.1
argparse==1.4.0
